Revert "Development"

+++ pkg/check/repo.go
@@ -1,9 +1,9 @@
 package check
 
 type Repo interface {
-	GetAllByRoundID(rID uint32) ([]*Check, error)
-	GetByRoundServiceID(rID uint32, sID uint32) (*Check, error)
-	Delete(rID uint32, sID uint32) error
+	GetAllByRoundID(rID uint64) ([]*Check, error)
+	GetByRoundServiceID(rID uint64, sID uint64) (*Check, error)
+	Delete(rID uint64, sID uint64) error
 	GetAll() ([]*Check, error)
 	Store(u *Check) error
 	StoreMany(u []*Check) error
