+++ pkg/check/serv.go
@@ -1,8 +1,8 @@
 package check
 
 type Serv interface {
-	GetAllByRoundID(rID uint32) ([]*Check, error)
-	GetByRoundServiceID(rID uint32, sID uint32) (*Check, error)
+	GetAllByRoundID(rID uint64) ([]*Check, error)
+	GetByRoundServiceID(rID uint64, sID uint64) (*Check, error)
 }
 
 type checkServ struct {
@@ -15,14 +15,14 @@ func NewCheckServ(repo Repo) Serv {
 	}
 }
 
-func (svc *checkServ) GetAllByRoundID(rID uint32) ([]*Check, error) {
+func (svc *checkServ) GetAllByRoundID(rID uint64) ([]*Check, error) {
 	return svc.repo.GetAllByRoundID(rID)
 }
-func (svc *checkServ) GetByRoundServiceID(rID uint32, sID uint32) (*Check, error) {
+func (svc *checkServ) GetByRoundServiceID(rID uint64, sID uint64) (*Check, error) {
 	return svc.repo.GetByRoundServiceID(rID, sID)
 }
 
-func (svc *checkServ) Delete(rID uint32, sID uint32) error { return svc.repo.Delete(rID, sID) }
+func (svc *checkServ) Delete(rID uint64, sID uint64) error { return svc.repo.Delete(rID, sID) }
 
 func (svc *checkServ) GetAll() ([]*Check, error) { return svc.repo.GetAll() }
