+++ pkg/host_group/serv.go
@@ -1,9 +1,9 @@
 package host_group
 
 type Serv interface {
-	Delete(id uint32) error
+	Delete(id uint64) error
 	GetAll() ([]*HostGroup, error)
-	GetByID(id uint32) (*HostGroup, error)
+	GetByID(id uint64) (*HostGroup, error)
 	Store(u *HostGroup) error
 	Update(u *HostGroup) error
 }
@@ -18,11 +18,11 @@ func NewHostGroupServ(repo Repo) Serv {
 	}
 }
 
-func (svc *hostGroupServ) Delete(id uint32) error { return svc.repo.Delete(id) }
+func (svc *hostGroupServ) Delete(id uint64) error { return svc.repo.Delete(id) }
 
 func (svc *hostGroupServ) GetAll() ([]*HostGroup, error) { return svc.repo.GetAll() }
 
-func (svc *hostGroupServ) GetByID(id uint32) (*HostGroup, error) { return svc.repo.GetByID(id) }
+func (svc *hostGroupServ) GetByID(id uint64) (*HostGroup, error) { return svc.repo.GetByID(id) }
 
 func (svc *hostGroupServ) Store(u *HostGroup) error { return svc.repo.Store(u) }
