+++ pkg/service/service.go
@@ -11,7 +11,7 @@ import (
 
 // Service Model represents a service that is being scored for a given host
 type Service struct {
-	ID uint32 `json:"id,omitempty" gorm:"primary_key"`
+	ID uint64 `json:"id,omitempty" gorm:"primary_key"`
 
 	//Name of the service that will be checked against known services
 	Name string `json:"name" gorm:"not null;default:null"`
@@ -19,21 +19,21 @@ type Service struct {
 	DisplayName string `json:"display_name,omitempty"`
 
 	// Points granted for a successful check
-	Points uint32 `json:"points" gorm:"not null;default: 0"`
+	Points uint64 `json:"points" gorm:"not null;default: 0"`
 
-	PointsBoost uint32 `json:"points_boost" gorm:"not null;default: 0"`
+	PointsBoost uint64 `json:"points_boost" gorm:"not null;default: 0"`
 
 	// The frequency of a service check. If round_units is 5 and round_delay is 0, then service checks will happen on every 5th round. (5,10, etc)
-	RoundUnits uint32 `json:"round_units,omitempty" gorm:"not null;default:1"`
+	RoundUnits uint64 `json:"round_units,omitempty" gorm:"not null;default:1"`
 
 	// The frequency of a service check. If round_units is 7 and round_delay is 3, then service checks will happen on every 7th round with an offset of 3. (10,17, etc)
-	RoundDelay *uint32 `json:"round_delay,omitempty" gorm:"not null;default: 0"`
+	RoundDelay *uint64 `json:"round_delay,omitempty" gorm:"not null;default: 0"`
 
 	// ID of a service group the service belongs to
-	ServiceGroupID uint32 `json:"service_group_id" gorm:"not null"`
+	ServiceGroupID uint64 `json:"service_group_id" gorm:"not null"`
 
 	// ID of a host the service belongs to
-	HostID uint32 `json:"host_id" gorm:"not null"`
+	HostID uint64 `json:"host_id" gorm:"not null"`
 
 	// Enables or Disables the service
 	Enabled *bool `json:"enabled,omitempty" gorm:"not null;default: true"`
