+++ pkg/team/repo.go
@@ -1,9 +1,9 @@
 package team
 
 type Repo interface {
-	Delete(id uint32) error
+	Delete(id uint64) error
 	GetAll() ([]*Team, error)
-	GetByID(id uint32) (*Team, error)
+	GetByID(id uint64) (*Team, error)
 	GetByName(name string) (*Team, error)
 	DeleteByName(name string) error
 	Store(u *Team) error
